Update level initialisation
Updated to assign level colors and created file groups

+++ SCIVisualiser/Model/TRIData.swift
@@ -11,5 +11,5 @@ import FirebaseFirestoreSwift
 internal struct TRIData: Identifiable, Codable {
     @DocumentID var id: String?
     let date: String
-    let tri: Float
+    let tri: Double
 }
